smarter changelog (#20138)
